+++ blocklets/ai-studio/src/components/file-editor/input/InputTable.tsx
@@ -25,6 +25,7 @@ import CursorTextIcon from '@iconify-icons/tabler/cursor-text';
 import DatabaseIcon from '@iconify-icons/tabler/database';
 import DotsIcon from '@iconify-icons/tabler/dots';
 import HistoryIcon from '@iconify-icons/tabler/history';
+import InfoCircleIcon from '@iconify-icons/tabler/info-circle';
 import MessageIcon from '@iconify-icons/tabler/message';
 import SquareNumberIcon from '@iconify-icons/tabler/square-number-1';
 import {
@@ -60,7 +61,7 @@ import { GridColDef } from '@mui/x-data-grid';
 import { useRequest } from 'ahooks';
 import { get, sortBy } from 'lodash';
 import { PopupState, bindDialog, bindPopper, bindTrigger, usePopupState } from 'material-ui-popup-state/hooks';
-import { useId, useMemo, useRef } from 'react';
+import { useId, useMemo, useRef, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { useAsync } from 'react-use';
 
@@ -993,6 +994,7 @@ function AgentParameter({ value, parameter }: { value: AssistantYjs; parameter: 
           <Typography variant="subtitle2">{t('chooseObject', { object: t('agent') })}</Typography>
 
           <AgentSelect
+            excludes={[value.id]}
             autoFocus
             disableClearable
             value={
@@ -1097,29 +1099,41 @@ function AuthorizeButton({ agent }: { agent: NonNullable<ReturnType<typeof useAg
   const dialogState = usePopupState({ variant: 'dialog' });
   const { projectId } = useCurrentProject();
 
+  const [authorized, setAuthorized] = useState(false);
   const { value: { secrets = [] } = {}, loading } = useAsync(() => getProjectInputSecrets(projectId), [projectId]);
 
-  const isAuthorized = useMemo(
-    () => secrets.find((i) => i.targetProjectId === agent.project.id && i.targetAgentId === agent.id),
-    [agent, secrets]
-  );
+  const isAuthorized =
+    useMemo(
+      () => secrets.find((i) => i.targetProjectId === agent.project.id && i.targetAgentId === agent.id),
+      [agent, secrets]
+    ) || authorized;
 
   if (!authInputs?.length || loading) return null;
 
   return (
-    <>
+    <Box>
       <Button variant={isAuthorized ? 'outlined' : 'contained'} fullWidth {...bindTrigger(dialogState)}>
         {t(isAuthorized ? 'reauthorize' : 'authorize')}
       </Button>
+      <Stack direction="row" alignItems="center" gap={0.5} my={0.5}>
+        <Box component={Icon} icon={InfoCircleIcon} color="text.secondary" />
+
+        <Typography variant="caption" sx={{ flex: 1 }}>
+          {t('authorizeApiKeyTip')}
+        </Typography>
+      </Stack>
 
       <AuthorizeParametersFormDialog
         agent={agent}
         maxWidth="sm"
         fullWidth
-        onSuccess={() => dialogState.close()}
+        onSuccess={() => {
+          setAuthorized(true);
+          dialogState.close();
+        }}
         {...bindDialog(dialogState)}
       />
-    </>
+    </Box>
   );
 }
 
@@ -1157,11 +1171,16 @@ function AuthorizeParametersFormDialog({
 
       <DialogContent>
         <Stack gap={1}>
-          {authInputs?.map((item) => (
+          {authInputs?.map((item, index) => (
             <Stack key={item.id}>
               <Typography variant="caption">{item.label || item.key}</Typography>
 
-              <PasswordField hiddenLabel fullWidth {...form.register(item.key!, { required: true })} />
+              <PasswordField
+                autoFocus={index === 0}
+                hiddenLabel
+                fullWidth
+                {...form.register(item.key!, { required: true })}
+              />
             </Stack>
           ))}
         </Stack>